rename: step1,2 진행시 사용했던 controller 이름 변경
