+++ tomcat/src/main/java/org/apache/coyote/http11/step1Controller/Step1Controller.java
@@ -1,10 +1,10 @@
-package org.apache.coyote.http11.controller;
+package org.apache.coyote.http11.step1Controller;
 
 import org.apache.coyote.http11.common.header.ContentTypeValue;
 import org.apache.coyote.http11.response.ResponseEntity;
 import org.apache.coyote.http11.response.statusLine.HttpStatus;
 
-public class Controller {
+public class Step1Controller {
 
     public ResponseEntity getIndex() {
         return new ResponseEntity(HttpStatus.OK, ContentTypeValue.TEXT_HTML, "Hello world!");
