+++ tomcat/src/main/java/org/apache/coyote/http11/step1Controller/Step1ControllerFinder.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package org.apache.coyote.http11.step1Controller;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.requestLine.HttpMethod;
@@ -6,33 +6,33 @@ import org.apache.coyote.http11.request.requestLine.requestUri.ResourcePath;
 import org.apache.coyote.http11.response.ResponseEntity;
 import org.apache.coyote.http11.response.statusLine.HttpStatus;
 
-public class ControllerFinder {
+public class Step1ControllerFinder {
 
-    private ControllerFinder() {
+    private Step1ControllerFinder() {
     }
 
     public static ResponseEntity find(final HttpRequest request) {
         final ResourcePath resourcePath = request.getResourcePath();
 
         if (resourcePath.isRootPath()) {
-            return new Controller().getIndex();
+            return new Step1Controller().getIndex();
         }
 
         if (request.isRequestOf(HttpMethod.POST)) {
             if (resourcePath.is("/login")) {
-                return new LoginController().postLogin(request.getRequestBody());
+                return new Step1LoginController().postLogin(request.getRequestBody());
             }
             if (resourcePath.is("/register")) {
-                return new RegisterController().postRegister(request.getRequestBody());
+                return new Step1RegisterController().postRegister(request.getRequestBody());
             }
         }
 
         if (request.isRequestOf(HttpMethod.GET)) {
             if (resourcePath.is("/login")) {
-                return new LoginController().getLogin(request);
+                return new Step1LoginController().getLogin(request);
             }
             if (resourcePath.is("/register")) {
-                return new RegisterController().getRegister();
+                return new Step1RegisterController().getRegister();
             }
         }
