+++ tomcat/src/test/java/org/apache/coyote/http11/controller/RegisterStep1ControllerTest.java
@@ -14,7 +14,7 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.util.List;
 
-class RegisterControllerTest {
+class RegisterStep1ControllerTest {
 
     @BeforeEach
     void beforeEach() {
